+++ package/kedro_viz/data_access/managers.py
@@ -14,7 +14,7 @@ from sqlalchemy.orm import sessionmaker
 
 from kedro_viz.constants import DEFAULT_REGISTERED_PIPELINE_ID, ROOT_MODULAR_PIPELINE_ID
 from kedro_viz.integrations.utils import UnavailableDataset
-from kedro_viz.models.flowchart import (
+from kedro_viz.models.flowchart.flowchart import (
     DataNode,
     GraphEdge,
     GraphNode,
