+++ package/kedro_viz/data_access/repositories/graph.py
@@ -3,7 +3,7 @@ centralise access to graph objects."""
 # pylint: disable=missing-class-docstring,missing-function-docstring
 from typing import Dict, Generator, List, Optional, Set
 
-from kedro_viz.models.flowchart import GraphEdge, GraphNode
+from kedro_viz.models.flowchart.flowchart import GraphEdge, GraphNode
 
 
 class GraphNodesRepository:
