+++ package/kedro_viz/data_access/repositories/tags.py
@@ -3,7 +3,7 @@ centralise access to tags data."""
 # pylint: disable=missing-class-docstring,missing-function-docstring
 from typing import Iterable, List, Set
 
-from kedro_viz.models.flowchart import Tag
+from kedro_viz.models.flowchart.flowchart import Tag
 
 
 class TagsRepository:
