+++ package/tests/test_data_access/test_managers.py
@@ -15,7 +15,7 @@ from kedro_viz.data_access.repositories.modular_pipelines import (
     ModularPipelinesRepository,
 )
 from kedro_viz.integrations.utils import UnavailableDataset
-from kedro_viz.models.flowchart import (
+from kedro_viz.models.flowchart.flowchart import (
     DataNode,
     GraphEdge,
     ParametersNode,
