+++ package/tests/test_data_access/test_repositories/test_graph.py
@@ -4,7 +4,7 @@ from kedro_viz.data_access.repositories import (
     GraphEdgesRepository,
     GraphNodesRepository,
 )
-from kedro_viz.models.flowchart import GraphEdge, GraphNode
+from kedro_viz.models.flowchart.flowchart import GraphEdge, GraphNode
 
 
 class TestGraphNodeRepository:
